requirements.txt
polars>=0.20.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
paho-mqtt>=1.6.0
apache-airflow>=2.8.1
python-dotenv>=1.0.0
numpy>=1.24.0
